changed some attribute of entities

+++ src/main/java/com/ss/email/registration/dto/AccountRegistrationRequest.java
@@ -1,15 +1,16 @@
 package com.ss.email.registration.dto;
 
+
 public class AccountRegistrationRequest  {
-    private final String email;
-    private final String firstName;
-    private final int account_id;
+    private final  String email;
+    private final  String firstName;
+    private final Long account_id;
 
     public String getEmail() {
         return email;
     }
 
-    public long getAccount_id() {
+    public Long getAccount_id() {
         return account_id;
     }
 
@@ -17,9 +18,10 @@ public class AccountRegistrationRequest  {
         return firstName;
     }
 
-    public AccountRegistrationRequest(String email, String firstName, int account_id) {
+    public AccountRegistrationRequest(String email, String firstName, Long account_id) {
         this.email = email;
         this.firstName = firstName;
         this.account_id = account_id;
     }
+
 }
