+++ src/main/java/com/ss/email/registration/dto/LoanRegistrationRequest.java
@@ -6,7 +6,7 @@ public class LoanRegistrationRequest {
     private final String firstName;
     private final long loan_id;
 
-    public LoanRegistrationRequest(String email, String firstName, int loan_id) {
+    public LoanRegistrationRequest(String email, String firstName, long loan_id) {
         this.email = email;
         this.firstName = firstName;
         this.loan_id = loan_id;
